requirements.txt
customtkinter
google-generativeai
python-dotenv
PyPDF2